Moves Thoughts endpoints to the dedicated directory
