+++ app/Http/Controllers/Api/Thoughts/CreateThoughtController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Thoughts;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreThoughtRequest;
