+++ app/Http/Controllers/Api/Thoughts/DeleteThoughtController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Thoughts;
 
 use App\Http\Controllers\Controller;
 use App\Models\Thought;
