+++ app/Http/Controllers/Api/Thoughts/ListingThoughtController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\Api\Thoughts;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ThoughtResource;
